refactor: move gstr3b details test to seperate file
